requirements.txt
Django==2.2.13
django-appconf==1.0.4
django-imagekit==4.0.2
pilkit==2.0
Pillow==7.2.0
pip==19.2.3
pytz==2020.1
setuptools==41.2.0
six==1.15.0
sqlparse==0.3.1
python-decouple==3.3